Integrates new DSL, Complete rewrite of the bootstrap class, Adds autoscale options.
Adds two new features to the OpenState class.  Provides the ability to add
required fields and default values for them.  Adds a new class instance variable
which returns all attributes that have been added to the OpenState instance.

Adds a new path method to the Config object which returns the path to the
directory containing the DSL definition files.  This should provide an easy way
for users to reference the config directory in the event they want to store
other non DSL (E.G. template) files in there and not use absolute paths.

The bootstrap class now simply reads the bootstrap object from the config class.
It then passes this binding to whatever template is provided.  This allows much
greater flexibility and removes logic about what should go in the user data from
the code.

Refactoring of the autoscale class to be less disgusting and adds in all currently
available configuration options.  It also provides "defaults" for many options
which are identical to the defaults AWS provides if you don't specify any.

Replaces options hash with an OpenStruct object.

Minor style changes like removing of parenthesis were not needed.

Bumps the version to 2.0.0