Uses new common interface for autoscaling groups.  Shifts normalization to the config class
This commit shifts responsbility for normalizing data into the config class.
Instead of forcing the data to always be normalized in the config class itself,
there is a config proxy class which will check to see if a normalize method
exists and call it if it does.

Adds a fix for tags not being updated if there were removals.